requirements.txt
# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles>=23.2.0
websockets>=12.0

# Machine Learning and OCR
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.45.0
accelerate==0.24.1
qwen-vl-utils
paddlepaddle>=2.5.0
paddleocr>=2.7.0

# Image processing
opencv-python-headless>=4.8.0
pillow==10.0.0
numpy>=1.24.0

# Utilities
requests>=2.31.0
jinja2>=3.1.0
pathlib2>=2.3.0
typing-extensions>=4.8.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
websocket-client>=1.6.0
